* Software and Emulator: Extra parameters property, it's a TStringList. Emulator can define optional parameters that was set in Software.
Example:

In (Q)ZDoom if you want to play a map that is not the first you can define an extra parameter: +map %0%
and in the software (map file) define an extra parameter: e2m2
%0% is the index of the line in ExtraParmeters software property. If empty whole parameter is no added. We can use %1%, %2%, etc.

In this case we can use "-warp %0%", but in software must be "2 2" but it's less intitutive.

This can be used in some MSX, CPC and Spectrum emulators as well, so we can autoload software.